LayoutTests/imported/w3c/web-platform-tests/css/css-transitions/animations/z-index-interpolation-expected.txt


PASS CSS Transitions: property <z-index> from neutral to [5] at (-0.3) should be [-4] 
PASS CSS Transitions: property <z-index> from neutral to [5] at (0) should be [-2] 
PASS CSS Transitions: property <z-index> from neutral to [5] at (0.3) should be [0] 
PASS CSS Transitions: property <z-index> from neutral to [5] at (0.6) should be [2] 
PASS CSS Transitions: property <z-index> from neutral to [5] at (1) should be [5] 
PASS CSS Transitions: property <z-index> from neutral to [5] at (1.5) should be [9] 
PASS CSS Transitions with transition: all: property <z-index> from neutral to [5] at (-0.3) should be [-4] 
PASS CSS Transitions with transition: all: property <z-index> from neutral to [5] at (0) should be [-2] 
PASS CSS Transitions with transition: all: property <z-index> from neutral to [5] at (0.3) should be [0] 
PASS CSS Transitions with transition: all: property <z-index> from neutral to [5] at (0.6) should be [2] 
PASS CSS Transitions with transition: all: property <z-index> from neutral to [5] at (1) should be [5] 
PASS CSS Transitions with transition: all: property <z-index> from neutral to [5] at (1.5) should be [9] 
PASS CSS Animations: property <z-index> from neutral to [5] at (-0.3) should be [-4] 
PASS CSS Animations: property <z-index> from neutral to [5] at (0) should be [-2] 
PASS CSS Animations: property <z-index> from neutral to [5] at (0.3) should be [0] 
PASS CSS Animations: property <z-index> from neutral to [5] at (0.6) should be [2] 
PASS CSS Animations: property <z-index> from neutral to [5] at (1) should be [5] 
PASS CSS Animations: property <z-index> from neutral to [5] at (1.5) should be [9] 
PASS Web Animations: property <z-index> from neutral to [5] at (-0.3) should be [-4] 
FAIL Web Animations: property <z-index> from neutral to [5] at (0) should be [-2] assert_equals: expected "- 2 " but got "- 4 "
FAIL Web Animations: property <z-index> from neutral to [5] at (0.3) should be [0] assert_equals: expected "0 " but got "- 1 "
FAIL Web Animations: property <z-index> from neutral to [5] at (0.6) should be [2] assert_equals: expected "2 " but got "3 "
PASS Web Animations: property <z-index> from neutral to [5] at (1) should be [5] 
FAIL Web Animations: property <z-index> from neutral to [5] at (1.5) should be [9] assert_equals: expected "9 " but got "5 "
FAIL CSS Transitions: property <z-index> from [initial] to [5] at (-0.3) should be [5] assert_equals: expected "5 " but got "- 1 "
FAIL CSS Transitions: property <z-index> from [initial] to [5] at (0) should be [5] assert_equals: expected "5 " but got "0 "
FAIL CSS Transitions: property <z-index> from [initial] to [5] at (0.3) should be [5] assert_equals: expected "5 " but got "2 "
FAIL CSS Transitions: property <z-index> from [initial] to [5] at (0.5) should be [5] assert_equals: expected "5 " but got "3 "
FAIL CSS Transitions: property <z-index> from [initial] to [5] at (0.6) should be [5] assert_equals: expected "5 " but got "3 "
PASS CSS Transitions: property <z-index> from [initial] to [5] at (1) should be [5] 
FAIL CSS Transitions: property <z-index> from [initial] to [5] at (1.5) should be [5] assert_equals: expected "5 " but got "8 "
FAIL CSS Transitions with transition: all: property <z-index> from [initial] to [5] at (-0.3) should be [5] assert_equals: expected "5 " but got "- 1 "
FAIL CSS Transitions with transition: all: property <z-index> from [initial] to [5] at (0) should be [5] assert_equals: expected "5 " but got "0 "
FAIL CSS Transitions with transition: all: property <z-index> from [initial] to [5] at (0.3) should be [5] assert_equals: expected "5 " but got "2 "
FAIL CSS Transitions with transition: all: property <z-index> from [initial] to [5] at (0.5) should be [5] assert_equals: expected "5 " but got "3 "
FAIL CSS Transitions with transition: all: property <z-index> from [initial] to [5] at (0.6) should be [5] assert_equals: expected "5 " but got "3 "
PASS CSS Transitions with transition: all: property <z-index> from [initial] to [5] at (1) should be [5] 
FAIL CSS Transitions with transition: all: property <z-index> from [initial] to [5] at (1.5) should be [5] assert_equals: expected "5 " but got "8 "
FAIL CSS Animations: property <z-index> from [initial] to [5] at (-0.3) should be [initial] assert_equals: expected "auto " but got "- 1 "
FAIL CSS Animations: property <z-index> from [initial] to [5] at (0) should be [initial] assert_equals: expected "auto " but got "0 "
FAIL CSS Animations: property <z-index> from [initial] to [5] at (0.3) should be [initial] assert_equals: expected "auto " but got "2 "
FAIL CSS Animations: property <z-index> from [initial] to [5] at (0.5) should be [5] assert_equals: expected "5 " but got "3 "
FAIL CSS Animations: property <z-index> from [initial] to [5] at (0.6) should be [5] assert_equals: expected "5 " but got "3 "
PASS CSS Animations: property <z-index> from [initial] to [5] at (1) should be [5] 
FAIL CSS Animations: property <z-index> from [initial] to [5] at (1.5) should be [5] assert_equals: expected "5 " but got "8 "
FAIL Web Animations: property <z-index> from [initial] to [5] at (-0.3) should be [initial] assert_equals: expected "auto " but got "- 1 "
FAIL Web Animations: property <z-index> from [initial] to [5] at (0) should be [initial] assert_equals: expected "auto " but got "0 "
FAIL Web Animations: property <z-index> from [initial] to [5] at (0.3) should be [initial] assert_equals: expected "auto " but got "2 "
FAIL Web Animations: property <z-index> from [initial] to [5] at (0.5) should be [5] assert_equals: expected "5 " but got "3 "
FAIL Web Animations: property <z-index> from [initial] to [5] at (0.6) should be [5] assert_equals: expected "5 " but got "3 "
PASS Web Animations: property <z-index> from [initial] to [5] at (1) should be [5] 
FAIL Web Animations: property <z-index> from [initial] to [5] at (1.5) should be [5] assert_equals: expected "5 " but got "8 "
PASS CSS Transitions: property <z-index> from [inherit] to [5] at (-0.3) should be [18] 
PASS CSS Transitions: property <z-index> from [inherit] to [5] at (0) should be [15] 
PASS CSS Transitions: property <z-index> from [inherit] to [5] at (0.3) should be [12] 
PASS CSS Transitions: property <z-index> from [inherit] to [5] at (0.6) should be [9] 
PASS CSS Transitions: property <z-index> from [inherit] to [5] at (1) should be [5] 
PASS CSS Transitions: property <z-index> from [inherit] to [5] at (1.5) should be [0] 
PASS CSS Transitions with transition: all: property <z-index> from [inherit] to [5] at (-0.3) should be [18] 
PASS CSS Transitions with transition: all: property <z-index> from [inherit] to [5] at (0) should be [15] 
PASS CSS Transitions with transition: all: property <z-index> from [inherit] to [5] at (0.3) should be [12] 
PASS CSS Transitions with transition: all: property <z-index> from [inherit] to [5] at (0.6) should be [9] 
PASS CSS Transitions with transition: all: property <z-index> from [inherit] to [5] at (1) should be [5] 
PASS CSS Transitions with transition: all: property <z-index> from [inherit] to [5] at (1.5) should be [0] 
PASS CSS Animations: property <z-index> from [inherit] to [5] at (-0.3) should be [18] 
PASS CSS Animations: property <z-index> from [inherit] to [5] at (0) should be [15] 
PASS CSS Animations: property <z-index> from [inherit] to [5] at (0.3) should be [12] 
PASS CSS Animations: property <z-index> from [inherit] to [5] at (0.6) should be [9] 
PASS CSS Animations: property <z-index> from [inherit] to [5] at (1) should be [5] 
PASS CSS Animations: property <z-index> from [inherit] to [5] at (1.5) should be [0] 
FAIL Web Animations: property <z-index> from [inherit] to [5] at (-0.3) should be [18] assert_equals: expected "18 " but got "- 4 "
FAIL Web Animations: property <z-index> from [inherit] to [5] at (0) should be [15] assert_equals: expected "15 " but got "- 4 "
FAIL Web Animations: property <z-index> from [inherit] to [5] at (0.3) should be [12] assert_equals: expected "12 " but got "- 1 "
FAIL Web Animations: property <z-index> from [inherit] to [5] at (0.6) should be [9] assert_equals: expected "9 " but got "3 "
PASS Web Animations: property <z-index> from [inherit] to [5] at (1) should be [5] 
FAIL Web Animations: property <z-index> from [inherit] to [5] at (1.5) should be [0] assert_equals: expected "0 " but got "5 "
FAIL CSS Transitions: property <z-index> from [unset] to [5] at (-0.3) should be [5] assert_equals: expected "5 " but got "- 1 "
FAIL CSS Transitions: property <z-index> from [unset] to [5] at (0) should be [5] assert_equals: expected "5 " but got "0 "
FAIL CSS Transitions: property <z-index> from [unset] to [5] at (0.3) should be [5] assert_equals: expected "5 " but got "2 "
FAIL CSS Transitions: property <z-index> from [unset] to [5] at (0.5) should be [5] assert_equals: expected "5 " but got "3 "
FAIL CSS Transitions: property <z-index> from [unset] to [5] at (0.6) should be [5] assert_equals: expected "5 " but got "3 "
PASS CSS Transitions: property <z-index> from [unset] to [5] at (1) should be [5] 
FAIL CSS Transitions: property <z-index> from [unset] to [5] at (1.5) should be [5] assert_equals: expected "5 " but got "8 "
FAIL CSS Transitions with transition: all: property <z-index> from [unset] to [5] at (-0.3) should be [5] assert_equals: expected "5 " but got "- 1 "
FAIL CSS Transitions with transition: all: property <z-index> from [unset] to [5] at (0) should be [5] assert_equals: expected "5 " but got "0 "
FAIL CSS Transitions with transition: all: property <z-index> from [unset] to [5] at (0.3) should be [5] assert_equals: expected "5 " but got "2 "
FAIL CSS Transitions with transition: all: property <z-index> from [unset] to [5] at (0.5) should be [5] assert_equals: expected "5 " but got "3 "
FAIL CSS Transitions with transition: all: property <z-index> from [unset] to [5] at (0.6) should be [5] assert_equals: expected "5 " but got "3 "
PASS CSS Transitions with transition: all: property <z-index> from [unset] to [5] at (1) should be [5] 
FAIL CSS Transitions with transition: all: property <z-index> from [unset] to [5] at (1.5) should be [5] assert_equals: expected "5 " but got "8 "
FAIL CSS Animations: property <z-index> from [unset] to [5] at (-0.3) should be [unset] assert_equals: expected "auto " but got "- 1 "
FAIL CSS Animations: property <z-index> from [unset] to [5] at (0) should be [unset] assert_equals: expected "auto " but got "0 "
FAIL CSS Animations: property <z-index> from [unset] to [5] at (0.3) should be [unset] assert_equals: expected "auto " but got "2 "
FAIL CSS Animations: property <z-index> from [unset] to [5] at (0.5) should be [5] assert_equals: expected "5 " but got "3 "
FAIL CSS Animations: property <z-index> from [unset] to [5] at (0.6) should be [5] assert_equals: expected "5 " but got "3 "
PASS CSS Animations: property <z-index> from [unset] to [5] at (1) should be [5] 
FAIL CSS Animations: property <z-index> from [unset] to [5] at (1.5) should be [5] assert_equals: expected "5 " but got "8 "
FAIL Web Animations: property <z-index> from [unset] to [5] at (-0.3) should be [unset] assert_equals: expected "auto " but got "- 1 "
FAIL Web Animations: property <z-index> from [unset] to [5] at (0) should be [unset] assert_equals: expected "auto " but got "0 "
FAIL Web Animations: property <z-index> from [unset] to [5] at (0.3) should be [unset] assert_equals: expected "auto " but got "2 "
FAIL Web Animations: property <z-index> from [unset] to [5] at (0.5) should be [5] assert_equals: expected "5 " but got "3 "
FAIL Web Animations: property <z-index> from [unset] to [5] at (0.6) should be [5] assert_equals: expected "5 " but got "3 "
PASS Web Animations: property <z-index> from [unset] to [5] at (1) should be [5] 
FAIL Web Animations: property <z-index> from [unset] to [5] at (1.5) should be [5] assert_equals: expected "5 " but got "8 "
PASS CSS Transitions: property <z-index> from [-5] to [5] at (-0.3) should be [-8] 
PASS CSS Transitions: property <z-index> from [-5] to [5] at (0) should be [-5] 
PASS CSS Transitions: property <z-index> from [-5] to [5] at (0.3) should be [-2] 
PASS CSS Transitions: property <z-index> from [-5] to [5] at (0.6) should be [1] 
PASS CSS Transitions: property <z-index> from [-5] to [5] at (1) should be [5] 
PASS CSS Transitions: property <z-index> from [-5] to [5] at (1.5) should be [10] 
PASS CSS Transitions with transition: all: property <z-index> from [-5] to [5] at (-0.3) should be [-8] 
PASS CSS Transitions with transition: all: property <z-index> from [-5] to [5] at (0) should be [-5] 
PASS CSS Transitions with transition: all: property <z-index> from [-5] to [5] at (0.3) should be [-2] 
PASS CSS Transitions with transition: all: property <z-index> from [-5] to [5] at (0.6) should be [1] 
PASS CSS Transitions with transition: all: property <z-index> from [-5] to [5] at (1) should be [5] 
PASS CSS Transitions with transition: all: property <z-index> from [-5] to [5] at (1.5) should be [10] 
PASS CSS Animations: property <z-index> from [-5] to [5] at (-0.3) should be [-8] 
PASS CSS Animations: property <z-index> from [-5] to [5] at (0) should be [-5] 
PASS CSS Animations: property <z-index> from [-5] to [5] at (0.3) should be [-2] 
PASS CSS Animations: property <z-index> from [-5] to [5] at (0.6) should be [1] 
PASS CSS Animations: property <z-index> from [-5] to [5] at (1) should be [5] 
PASS CSS Animations: property <z-index> from [-5] to [5] at (1.5) should be [10] 
PASS Web Animations: property <z-index> from [-5] to [5] at (-0.3) should be [-8] 
PASS Web Animations: property <z-index> from [-5] to [5] at (0) should be [-5] 
PASS Web Animations: property <z-index> from [-5] to [5] at (0.3) should be [-2] 
PASS Web Animations: property <z-index> from [-5] to [5] at (0.6) should be [1] 
PASS Web Animations: property <z-index> from [-5] to [5] at (1) should be [5] 
PASS Web Animations: property <z-index> from [-5] to [5] at (1.5) should be [10] 
PASS CSS Transitions: property <z-index> from [2] to [4] at (-0.3) should be [1] 
PASS CSS Transitions: property <z-index> from [2] to [4] at (0) should be [2] 
PASS CSS Transitions: property <z-index> from [2] to [4] at (0.3) should be [3] 
PASS CSS Transitions: property <z-index> from [2] to [4] at (0.6) should be [3] 
PASS CSS Transitions: property <z-index> from [2] to [4] at (1) should be [4] 
PASS CSS Transitions: property <z-index> from [2] to [4] at (1.5) should be [5] 
PASS CSS Transitions with transition: all: property <z-index> from [2] to [4] at (-0.3) should be [1] 
PASS CSS Transitions with transition: all: property <z-index> from [2] to [4] at (0) should be [2] 
PASS CSS Transitions with transition: all: property <z-index> from [2] to [4] at (0.3) should be [3] 
PASS CSS Transitions with transition: all: property <z-index> from [2] to [4] at (0.6) should be [3] 
PASS CSS Transitions with transition: all: property <z-index> from [2] to [4] at (1) should be [4] 
PASS CSS Transitions with transition: all: property <z-index> from [2] to [4] at (1.5) should be [5] 
PASS CSS Animations: property <z-index> from [2] to [4] at (-0.3) should be [1] 
PASS CSS Animations: property <z-index> from [2] to [4] at (0) should be [2] 
PASS CSS Animations: property <z-index> from [2] to [4] at (0.3) should be [3] 
PASS CSS Animations: property <z-index> from [2] to [4] at (0.6) should be [3] 
PASS CSS Animations: property <z-index> from [2] to [4] at (1) should be [4] 
PASS CSS Animations: property <z-index> from [2] to [4] at (1.5) should be [5] 
PASS Web Animations: property <z-index> from [2] to [4] at (-0.3) should be [1] 
PASS Web Animations: property <z-index> from [2] to [4] at (0) should be [2] 
PASS Web Animations: property <z-index> from [2] to [4] at (0.3) should be [3] 
PASS Web Animations: property <z-index> from [2] to [4] at (0.6) should be [3] 
PASS Web Animations: property <z-index> from [2] to [4] at (1) should be [4] 
PASS Web Animations: property <z-index> from [2] to [4] at (1.5) should be [5] 
PASS CSS Transitions: property <z-index> from [-2] to [-4] at (-0.3) should be [-1] 
PASS CSS Transitions: property <z-index> from [-2] to [-4] at (0) should be [-2] 
PASS CSS Transitions: property <z-index> from [-2] to [-4] at (0.1) should be [-2] 
PASS CSS Transitions: property <z-index> from [-2] to [-4] at (0.3) should be [-3] 
PASS CSS Transitions: property <z-index> from [-2] to [-4] at (0.6) should be [-3] 
PASS CSS Transitions: property <z-index> from [-2] to [-4] at (1) should be [-4] 
PASS CSS Transitions: property <z-index> from [-2] to [-4] at (1.5) should be [-5] 
PASS CSS Transitions with transition: all: property <z-index> from [-2] to [-4] at (-0.3) should be [-1] 
PASS CSS Transitions with transition: all: property <z-index> from [-2] to [-4] at (0) should be [-2] 
PASS CSS Transitions with transition: all: property <z-index> from [-2] to [-4] at (0.1) should be [-2] 
PASS CSS Transitions with transition: all: property <z-index> from [-2] to [-4] at (0.3) should be [-3] 
PASS CSS Transitions with transition: all: property <z-index> from [-2] to [-4] at (0.6) should be [-3] 
PASS CSS Transitions with transition: all: property <z-index> from [-2] to [-4] at (1) should be [-4] 
PASS CSS Transitions with transition: all: property <z-index> from [-2] to [-4] at (1.5) should be [-5] 
PASS CSS Animations: property <z-index> from [-2] to [-4] at (-0.3) should be [-1] 
PASS CSS Animations: property <z-index> from [-2] to [-4] at (0) should be [-2] 
PASS CSS Animations: property <z-index> from [-2] to [-4] at (0.1) should be [-2] 
PASS CSS Animations: property <z-index> from [-2] to [-4] at (0.3) should be [-3] 
PASS CSS Animations: property <z-index> from [-2] to [-4] at (0.6) should be [-3] 
PASS CSS Animations: property <z-index> from [-2] to [-4] at (1) should be [-4] 
PASS CSS Animations: property <z-index> from [-2] to [-4] at (1.5) should be [-5] 
PASS Web Animations: property <z-index> from [-2] to [-4] at (-0.3) should be [-1] 
PASS Web Animations: property <z-index> from [-2] to [-4] at (0) should be [-2] 
PASS Web Animations: property <z-index> from [-2] to [-4] at (0.1) should be [-2] 
PASS Web Animations: property <z-index> from [-2] to [-4] at (0.3) should be [-3] 
PASS Web Animations: property <z-index> from [-2] to [-4] at (0.6) should be [-3] 
PASS Web Animations: property <z-index> from [-2] to [-4] at (1) should be [-4] 
PASS Web Animations: property <z-index> from [-2] to [-4] at (1.5) should be [-5] 
FAIL CSS Transitions: property <z-index> from [auto] to [10] at (-0.3) should be [10] assert_equals: expected "10 " but got "- 3 "
FAIL CSS Transitions: property <z-index> from [auto] to [10] at (0) should be [10] assert_equals: expected "10 " but got "0 "
FAIL CSS Transitions: property <z-index> from [auto] to [10] at (0.3) should be [10] assert_equals: expected "10 " but got "3 "
FAIL CSS Transitions: property <z-index> from [auto] to [10] at (0.5) should be [10] assert_equals: expected "10 " but got "5 "
FAIL CSS Transitions: property <z-index> from [auto] to [10] at (0.6) should be [10] assert_equals: expected "10 " but got "6 "
PASS CSS Transitions: property <z-index> from [auto] to [10] at (1) should be [10] 
FAIL CSS Transitions: property <z-index> from [auto] to [10] at (1.5) should be [10] assert_equals: expected "10 " but got "15 "
FAIL CSS Transitions with transition: all: property <z-index> from [auto] to [10] at (-0.3) should be [10] assert_equals: expected "10 " but got "- 3 "
FAIL CSS Transitions with transition: all: property <z-index> from [auto] to [10] at (0) should be [10] assert_equals: expected "10 " but got "0 "
FAIL CSS Transitions with transition: all: property <z-index> from [auto] to [10] at (0.3) should be [10] assert_equals: expected "10 " but got "3 "
FAIL CSS Transitions with transition: all: property <z-index> from [auto] to [10] at (0.5) should be [10] assert_equals: expected "10 " but got "5 "
FAIL CSS Transitions with transition: all: property <z-index> from [auto] to [10] at (0.6) should be [10] assert_equals: expected "10 " but got "6 "
PASS CSS Transitions with transition: all: property <z-index> from [auto] to [10] at (1) should be [10] 
FAIL CSS Transitions with transition: all: property <z-index> from [auto] to [10] at (1.5) should be [10] assert_equals: expected "10 " but got "15 "
FAIL CSS Animations: property <z-index> from [auto] to [10] at (-0.3) should be [auto] assert_equals: expected "auto " but got "- 3 "
FAIL CSS Animations: property <z-index> from [auto] to [10] at (0) should be [auto] assert_equals: expected "auto " but got "0 "
FAIL CSS Animations: property <z-index> from [auto] to [10] at (0.3) should be [auto] assert_equals: expected "auto " but got "3 "
FAIL CSS Animations: property <z-index> from [auto] to [10] at (0.5) should be [10] assert_equals: expected "10 " but got "5 "
FAIL CSS Animations: property <z-index> from [auto] to [10] at (0.6) should be [10] assert_equals: expected "10 " but got "6 "
PASS CSS Animations: property <z-index> from [auto] to [10] at (1) should be [10] 
FAIL CSS Animations: property <z-index> from [auto] to [10] at (1.5) should be [10] assert_equals: expected "10 " but got "15 "
FAIL Web Animations: property <z-index> from [auto] to [10] at (-0.3) should be [auto] assert_equals: expected "auto " but got "- 3 "
FAIL Web Animations: property <z-index> from [auto] to [10] at (0) should be [auto] assert_equals: expected "auto " but got "0 "
FAIL Web Animations: property <z-index> from [auto] to [10] at (0.3) should be [auto] assert_equals: expected "auto " but got "3 "
FAIL Web Animations: property <z-index> from [auto] to [10] at (0.5) should be [10] assert_equals: expected "10 " but got "5 "
FAIL Web Animations: property <z-index> from [auto] to [10] at (0.6) should be [10] assert_equals: expected "10 " but got "6 "
PASS Web Animations: property <z-index> from [auto] to [10] at (1) should be [10] 
FAIL Web Animations: property <z-index> from [auto] to [10] at (1.5) should be [10] assert_equals: expected "10 " but got "15 "
